feat(filterchecks): add optional metrics struct for caching
Add an optional metrics struct that if present, keeps track of the
efficacy of the extraction cache/evaluation cache.

The metrics count the number of times extract_cached() was called and
when it could use the cache, and the number of times comapre() was
called and when it could use the cache.

Like the caches, the memory for the pointer is assumed to be owned by
something else.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>